+++ tests/integration/test_iivsearch.py
@@ -10,7 +10,9 @@ from pharmpy.tools import fit, retrieve_models, run_iivsearch
 
 def test_block_structure(tmp_path, model_count, start_model):
     with chdir(tmp_path):
-        res = run_iivsearch('brute_force_block_structure', model=start_model)
+        res = run_iivsearch(
+            'brute_force_block_structure', results=start_model.modelfit_results, model=start_model
+        )
 
         no_of_candidate_models = 4
         assert len(res.summary_tool) == no_of_candidate_models + 1
@@ -44,7 +46,9 @@ def test_block_structure(tmp_path, model_count, start_model):
 
 def test_no_of_etas(tmp_path, model_count, start_model):
     with chdir(tmp_path):
-        res = run_iivsearch('brute_force_no_of_etas', model=start_model)
+        res = run_iivsearch(
+            'brute_force_no_of_etas', results=start_model.modelfit_results, model=start_model
+        )
 
         no_of_candidate_models = 7
         assert len(res.summary_tool) == no_of_candidate_models + 1
@@ -74,7 +78,7 @@ def test_no_of_etas(tmp_path, model_count, start_model):
 
 def test_brute_force(tmp_path, model_count, start_model):
     with chdir(tmp_path):
-        res = run_iivsearch('brute_force', model=start_model)
+        res = run_iivsearch('brute_force', results=start_model.modelfit_results, model=start_model)
 
         no_of_candidate_models = 8
         assert len(res.summary_tool) == no_of_candidate_models + 2
@@ -136,6 +140,7 @@ def test_no_of_etas_iiv_strategies(tmp_path, model_count, start_model, iiv_strat
         res = run_iivsearch(
             'brute_force_no_of_etas',
             iiv_strategy=iiv_strategy,
+            results=start_model.modelfit_results,
             model=start_model,
         )